pdf-question-answering-rag/requirements.txt
streamlit
pypdf
sentence-transformers
transformers
torch
faiss-cpu
scikit-learn
numpy
pickle-mixin
